+++ apps/web/src/features/moim-create/index.ts
@@ -1,2 +1,3 @@
+export { parseMoimFormData } from "./model/parse-moim-form-data";
 export { MoimCreateForm } from "./ui/moim-create-form";
-export { ThemeColorSelect } from "./ui/theme-color-select";
+export { MoimFormFields } from "./ui/moim-form-fields";
